removed handler dependence from transactions module
